+++ SkillAPI/src/com/sucy/skill/mechanic/PotionMechanic.java
@@ -38,8 +38,8 @@ public class PotionMechanic implements IMechanic {
         // Get attributes
         int level = data.getSkillLevel(skill.getName());
         PotionEffectType potionType = PotionEffectType.values()[skill.getValue(TYPE)];
-        int duration = skill.getAttribute(DURATION, target, level) * 20;
-        int tier = skill.getAttribute(TIER, target, level);
+        int duration = (int)(skill.getAttribute(DURATION, target, level) * 20);
+        int tier = (int)skill.getAttribute(TIER, target, level);
 
         // Must have a target
         if (targets.size() == 0) return false;
